Split CLI merge module into discovery and diag sub-modules
`src/cli/merge.rs` exceeded the 400-line limit. Extract file-layer
discovery and loading into `src/cli/discovery.rs` and diagnostic-JSON
resolution helpers into `src/cli/diag.rs`, keeping the merge pipeline
logic in `merge.rs`.

`env_provider` is now `pub(crate)` so the diag module can call it.
`resolve_merged_diag_json` remains the public entry point, re-exported
from `cli::diag` through `cli::mod`.
