+++ src/cli/merge.rs
@@ -5,8 +5,7 @@
 //! It implements the full four-layer merge pipeline:
 //!
 //! 1. **Defaults** — `CliConfig::default()` serialised to JSON.
-//! 2. **File layers** — discovered via `NETSUKE_CONFIG_PATH`, `NETSUKE_CONFIG`,
-//!    or directory walk; loaded through `OrthoConfig`'s `ConfigDiscovery`.
+//! 2. **File layers** — discovered and loaded by [`super::discovery`].
 //! 3. **Environment layer** — `NETSUKE_`-prefixed variables normalised via
 //!    `Uncased` and merged through Figment.
 //! 4. **CLI override layer** — fields explicitly supplied on the command line
@@ -18,31 +17,25 @@
 //! - Applies `CliConfig`'s `PostMergeHook` for cross-field validation.
 //! - Produces a fully resolved `Cli` for the runner.
 //!
-//! The secondary entry point, [`resolve_merged_diag_json`], performs a
-//! lightweight pre-merge pass to determine diagnostic-output mode before the
-//! full merge runs.
+//! Diagnostic JSON resolution lives in [`super::diag`] so it can run before
+//! the full merge.
 
 use clap::ArgMatches;
 use clap::parser::ValueSource;
 use ortho_config::declarative::LayerComposition;
 use ortho_config::figment::{Figment, providers::Env};
 use ortho_config::uncased::Uncased;
-use ortho_config::{ConfigDiscovery, MergeComposer, OrthoMergeExt, OrthoResult, sanitize_value};
-use ortho_config::{MergeLayer, load_config_file_as_chain};
+use ortho_config::{MergeComposer, OrthoMergeExt, OrthoResult, sanitize_value};
 use serde::Serialize;
-use std::borrow::Cow;
-use std::io;
-use std::path::{Path, PathBuf};
-use std::sync::Arc;
 
 use serde_json::{Map, Value, json};
 
 use super::config::{BuildConfig, CliConfig, Theme};
+use super::discovery::push_file_layers;
 use super::parser::{BuildArgs, Cli, Commands};
 use super::validation_error;
 use crate::theme::ThemePreference;
-const CONFIG_ENV_VAR: &str = "NETSUKE_CONFIG";
-const CONFIG_ENV_VAR_LEGACY: &str = "NETSUKE_CONFIG_PATH";
+
 const ENV_PREFIX: &str = "NETSUKE_";
 
 /// Merge discovered configuration layers over parsed CLI input.
@@ -85,171 +78,12 @@ pub fn merge_with_config(cli: &Cli, matches: &ArgMatches) -> OrthoResult<Cli> {
     Ok(apply_config(cli, merged))
 }
 
-fn push_file_layers(
-    cli: &Cli,
-    composer: &mut MergeComposer,
-    errors: &mut Vec<Arc<ortho_config::OrthoError>>,
-) {
-    let layers_result = explicit_config_path(cli).map_or_else(
-        || collect_file_layers(cli.directory.as_deref()),
-        |path| load_layers_from_path(&path),
-    );
-    match layers_result {
-        Ok(layers) => {
-            for layer in layers {
-                composer.push_layer(layer);
-            }
-        }
-        Err(err) => errors.push(err),
-    }
-}
-
-fn env_provider() -> Env {
+pub(crate) fn env_provider() -> Env {
     Env::prefixed(ENV_PREFIX)
-}
-
-fn config_discovery(directory: Option<&PathBuf>) -> ConfigDiscovery {
-    let mut builder = ConfigDiscovery::builder("netsuke").env_var(CONFIG_ENV_VAR_LEGACY);
-    if let Some(dir) = directory {
-        builder = builder.clear_project_roots().add_project_root(dir);
-    }
-    builder.build()
-}
-
-fn collect_file_layers(directory: Option<&Path>) -> OrthoResult<Vec<MergeLayer<'static>>> {
-    let discovery = config_discovery(directory.map(PathBuf::from).as_ref());
-    let mut file_layers = discovery.compose_layers();
-    let mut errors = file_layers.required_errors;
-    if file_layers.value.is_empty() {
-        errors.append(&mut file_layers.optional_errors);
-    }
-    if let Some(err) = errors.into_iter().next() {
-        return Err(err);
-    }
-
-    let project_file = project_scope_file_str(directory);
-    let has_project_layer = file_layers.value.iter().any(|layer| {
-        layer
-            .path()
-            .is_some_and(|path| project_file.as_deref() == Some(path.as_str()))
-    });
-    if has_project_layer {
-        return Ok(file_layers.value);
-    }
-
-    let project_layers = project_scope_layers(directory)?;
-    Ok(file_layers
-        .value
-        .into_iter()
-        .chain(project_layers)
-        .collect())
-}
-
-fn project_scope_file_str(directory: Option<&Path>) -> Option<String> {
-    let root = directory
-        .map(PathBuf::from)
-        .or_else(|| std::env::current_dir().ok())?;
-    root.join(".netsuke.toml").to_str().map(String::from)
-}
-
-fn project_scope_layers(directory: Option<&Path>) -> OrthoResult<Vec<MergeLayer<'static>>> {
-    let root = directory
-        .map(PathBuf::from)
-        .or_else(|| std::env::current_dir().ok());
-    let Some(project_file) = root.map(|dir| dir.join(".netsuke.toml")) else {
-        return Ok(Vec::new());
-    };
-    match load_config_file_as_chain(&project_file) {
-        Ok(Some(chain)) => Ok(chain
-            .values
-            .into_iter()
-            .map(|(value, path)| MergeLayer::file(Cow::Owned(value), Some(path)))
-            .collect()),
-        Ok(None) => Ok(Vec::new()),
-        Err(err) => Err(err),
-    }
-}
-
-fn explicit_config_path(cli: &Cli) -> Option<PathBuf> {
-    cli.config
-        .clone()
-        .or_else(|| env_config_path(CONFIG_ENV_VAR))
-        .or_else(|| env_config_path(CONFIG_ENV_VAR_LEGACY))
-}
-
-fn env_config_path(var_name: &str) -> Option<PathBuf> {
-    std::env::var_os(var_name)
-        .filter(|value| !value.is_empty())
-        .map(PathBuf::from)
-}
-
-fn load_layers_from_path(path: &std::path::Path) -> OrthoResult<Vec<MergeLayer<'static>>> {
-    match load_config_file_as_chain(path) {
-        Ok(Some(chain)) => Ok(chain
-            .values
-            .into_iter()
-            .map(|(value, layer_path)| MergeLayer::file(Cow::Owned(value), Some(layer_path)))
-            .collect()),
-        Ok(None) => Err(Arc::new(ortho_config::OrthoError::File {
-            path: path.to_path_buf(),
-            source: Box::new(io::Error::new(
-                io::ErrorKind::NotFound,
-                "explicit configuration file not found",
-            )),
-        })),
-        Err(err) => Err(err),
-    }
 }
 
 fn is_empty_value(value: &Value) -> bool {
     matches!(value, Value::Object(map) if map.is_empty())
-}
-
-fn diag_json_from_layer(value: &Value) -> Option<bool> {
-    value
-        .as_object()
-        .and_then(|map| map.get("diag_json"))
-        .and_then(Value::as_bool)
-}
-
-fn collect_diag_file_layers(cli: &Cli) -> OrthoResult<Vec<MergeLayer<'static>>> {
-    explicit_config_path(cli).map_or_else(
-        || collect_file_layers(cli.directory.as_deref()),
-        |path| load_layers_from_path(&path),
-    )
-}
-
-fn diag_json_from_matches(cli: &Cli, matches: &ArgMatches, discovered: bool) -> bool {
-    if matches.value_source("output_format") == Some(ValueSource::CommandLine) {
-        cli.resolved_diag_json()
-    } else if matches.value_source("diag_json") == Some(ValueSource::CommandLine) {
-        cli.diag_json
-    } else {
-        discovered
-    }
-}
-
-fn diag_json_from_file_layers(cli: &Cli) -> OrthoResult<bool> {
-    let default = Cli::default().diag_json;
-    let layers = collect_diag_file_layers(cli)?;
-    let mut diag_json = default;
-    for layer in layers {
-        if let Some(layer_diag_json) = diag_json_from_layer(&layer.into_value()) {
-            diag_json = layer_diag_json;
-        }
-    }
-    Ok(diag_json)
-}
-
-fn diag_json_from_env(fallback: bool) -> bool {
-    let env_provider = env_provider()
-        .map(|key| Uncased::new(key.as_str().to_ascii_uppercase()))
-        .split("__");
-    Figment::from(env_provider)
-        .extract::<serde_json::Value>()
-        .ok()
-        .and_then(|value| diag_json_from_layer(&value))
-        .unwrap_or(fallback)
 }
 
 fn validate_output_format_source(config: &CliConfig, matches: &ArgMatches) -> OrthoResult<()> {
@@ -262,19 +96,6 @@ fn validate_output_format_source(config: &CliConfig, matches: &ArgMatches) -> Or
         ));
     }
     Ok(())
-}
-
-/// Resolve the effective diagnostic JSON preference from the raw config layers.
-///
-/// This is used before full config merging so startup and merge-time failures
-/// can still honour `diag_json` values sourced from config files or the
-/// environment.
-#[must_use]
-pub fn resolve_merged_diag_json(cli: &Cli, matches: &ArgMatches) -> bool {
-    let mut diag_json =
-        diag_json_from_file_layers(cli).unwrap_or_else(|_| Cli::default().diag_json);
-    diag_json = diag_json_from_env(diag_json);
-    diag_json_from_matches(cli, matches, diag_json)
 }
 
 fn cli_overrides_from_matches(cli: &Cli, matches: &ArgMatches) -> OrthoResult<Value> {
